refactor: move dashboard components to shared `/dashboard` route for improved modularity
note: profile settings inputs and toggles are placeholders and do not have functionality set up as yet
