+++ app/dashboard/companies/_components/CompanyList.tsx
@@ -3,9 +3,9 @@
 import { useCallback, useMemo, useState } from "react";
 
 import { useCompaniesQuery } from "@/actions/queries/admin.queries";
-import CompanyManipulations from "@/app/app/companies/_components/CompanyManipulations";
-import DeleteCompany from "@/app/app/companies/_components/DeleteCompany";
-import Skeleton from "@/app/app/companies/_components/Skeleton";
+import CompanyManipulations from "@/app/dashboard/companies/_components/CompanyManipulations";
+import DeleteCompany from "@/app/dashboard/companies/_components/DeleteCompany";
+import Skeleton from "@/app/dashboard/companies/_components/Skeleton";
 import { Prisma } from "@prisma/client";
 import { formatDistanceToNow } from "date-fns";
 import { MoreHorizontal } from "lucide-react";
@@ -62,7 +62,7 @@ const CompanyTable = ({ sessionUser }: { sessionUser: User }) => {
   if (!companies) return <NoDataFallback />;
 
   return (
-    <div className="bg-background/30 border rounded-md">
+    <div className="rounded-md border bg-background/30">
       <Table>
         <TableHeader>
           <TableRow>
@@ -162,9 +162,9 @@ const CompanyContextMenu = ({
     <>
       <DropdownMenu>
         <DropdownMenuTrigger asChild>
-          <Button variant="ghost" className="p-0 w-8 h-8">
+          <Button variant="ghost" className="h-8 w-8 p-0">
             <span className="sr-only">Open menu</span>
-            <MoreHorizontal className="w-4 h-4" />
+            <MoreHorizontal className="h-4 w-4" />
           </Button>
         </DropdownMenuTrigger>
         <DropdownMenuContent align="end">
