+++ app/dashboard/companies/_components/DeleteCompany.tsx
