+++ app/dashboard/cvs/all/_components/DocList.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { useAdminDocsQuery } from "@/actions/queries/admin.queries";
-import DocContextMenu from "@/app/app/cvs/_components/ContextMenu";
-import Skeleton from "@/app/app/cvs/_components/Skeleton";
+import DocContextMenu from "@/app/dashboard/cvs/_components/ContextMenu";
+import Skeleton from "@/app/dashboard/cvs/_components/Skeleton";
 import { formatDistanceToNow } from "date-fns";
 import { User } from "next-auth";
