+++ app/dashboard/layout.tsx
@@ -1,15 +1,16 @@
+import { DashboardLayout } from "@/app/dashboard/_components/DashboardLayout";
 import { ErrorBoundary } from "react-error-boundary";
 
 import FallBack from "@/components/global/Fallback";
-import Footer from "@/components/global/Footer";
-import Header from "@/components/global/Header";
+
+import { requireAuth } from "@/lib/utils";
 
 async function layout({ children }: { children: React.ReactNode }) {
+  const { user } = await requireAuth("/dashboard");
+
   return (
     <ErrorBoundary FallbackComponent={FallBack}>
-      <Header />
-      <section className="layout">{children}</section>
-      <Footer />
+      <DashboardLayout user={user}>{children}</DashboardLayout>
     </ErrorBoundary>
   );
 }
