+++ app/dashboard/templates/_components/ViewTemplate.tsx
